Fix #3349 : Changes suggested by code review. Add an option for the single request parameter.

+++ samples/client/petstore/typescript-fetch/builds/with-interfaces/apis/StoreApi.ts
@@ -19,6 +19,18 @@ import {
     OrderToJSON,
 } from '../models';
 
+export interface DeleteOrderRequest {
+    orderId: string;
+}
+
+export interface GetOrderByIdRequest {
+    orderId: number;
+}
+
+export interface PlaceOrderRequest {
+    body: Order;
+}
+
 /**
  * no description
  */
@@ -28,13 +40,17 @@ export class StoreApi extends runtime.BaseAPI {
      * For valid response try integer IDs with value < 1000. Anything above 1000 or nonintegers will generate API errors
      * Delete purchase order by ID
      */
-    async deleteOrderRaw(orderId: string): Promise<runtime.ApiResponse<void>> {
+    async deleteOrderRaw(requestParameters: DeleteOrderRequest): Promise<runtime.ApiResponse<void>> {
+        if (requestParameters.orderId === null || requestParameters.orderId === undefined) {
+            throw new runtime.RequiredError('orderId','Required parameter requestParameters.orderId was null or undefined when calling deleteOrder.');
+        }
+
         const queryParameters: runtime.HTTPQuery = {};
 
         const headerParameters: runtime.HTTPHeaders = {};
 
         const response = await this.request({
-            path: `/store/order/{orderId}`.replace(`{${"orderId"}}`, encodeURIComponent(String(orderId))),
+            path: `/store/order/{orderId}`.replace(`{${"orderId"}}`, encodeURIComponent(String(requestParameters.orderId))),
             method: 'DELETE',
             headers: headerParameters,
             query: queryParameters,
@@ -43,12 +59,12 @@ export class StoreApi extends runtime.BaseAPI {
         return new runtime.VoidApiResponse(response);
     }
 
-    /**
-     * For valid response try integer IDs with value < 1000. Anything above 1000 or nonintegers will generate API errors
-     * Delete purchase order by ID
-     */
+   /**
+    * For valid response try integer IDs with value < 1000. Anything above 1000 or nonintegers will generate API errors
+    * Delete purchase order by ID
+    */
     async deleteOrder(orderId: string): Promise<void> {
-        await this.deleteOrderRaw(orderId);
+        await this.deleteOrderRaw({ orderId: orderId });
     }
 
     /**
@@ -74,12 +90,12 @@ export class StoreApi extends runtime.BaseAPI {
         return new runtime.JSONApiResponse<any>(response);
     }
 
-    /**
-     * Returns a map of status codes to quantities
-     * Returns pet inventories by status
-     */
+   /**
+    * Returns a map of status codes to quantities
+    * Returns pet inventories by status
+    */
     async getInventory(): Promise<{ [key: string]: number; }> {
-        const response = await this.getInventoryRaw();
+        const response = await this.getInventoryRaw({  });
         return await response.value();
     }
 
@@ -87,13 +103,17 @@ export class StoreApi extends runtime.BaseAPI {
      * For valid response try integer IDs with value <= 5 or > 10. Other values will generated exceptions
      * Find purchase order by ID
      */
-    async getOrderByIdRaw(orderId: number): Promise<runtime.ApiResponse<Order>> {
+    async getOrderByIdRaw(requestParameters: GetOrderByIdRequest): Promise<runtime.ApiResponse<Order>> {
+        if (requestParameters.orderId === null || requestParameters.orderId === undefined) {
+            throw new runtime.RequiredError('orderId','Required parameter requestParameters.orderId was null or undefined when calling getOrderById.');
+        }
+
         const queryParameters: runtime.HTTPQuery = {};
 
         const headerParameters: runtime.HTTPHeaders = {};
 
         const response = await this.request({
-            path: `/store/order/{orderId}`.replace(`{${"orderId"}}`, encodeURIComponent(String(orderId))),
+            path: `/store/order/{orderId}`.replace(`{${"orderId"}}`, encodeURIComponent(String(requestParameters.orderId))),
             method: 'GET',
             headers: headerParameters,
             query: queryParameters,
@@ -102,19 +122,23 @@ export class StoreApi extends runtime.BaseAPI {
         return new runtime.JSONApiResponse(response, (jsonValue) => OrderFromJSON(jsonValue));
     }
 
-    /**
-     * For valid response try integer IDs with value <= 5 or > 10. Other values will generated exceptions
-     * Find purchase order by ID
-     */
+   /**
+    * For valid response try integer IDs with value <= 5 or > 10. Other values will generated exceptions
+    * Find purchase order by ID
+    */
     async getOrderById(orderId: number): Promise<Order> {
-        const response = await this.getOrderByIdRaw(orderId);
+        const response = await this.getOrderByIdRaw({ orderId: orderId });
         return await response.value();
     }
 
     /**
      * Place an order for a pet
      */
-    async placeOrderRaw(body: Order): Promise<runtime.ApiResponse<Order>> {
+    async placeOrderRaw(requestParameters: PlaceOrderRequest): Promise<runtime.ApiResponse<Order>> {
+        if (requestParameters.body === null || requestParameters.body === undefined) {
+            throw new runtime.RequiredError('body','Required parameter requestParameters.body was null or undefined when calling placeOrder.');
+        }
+
         const queryParameters: runtime.HTTPQuery = {};
 
         const headerParameters: runtime.HTTPHeaders = {};
@@ -126,17 +150,17 @@ export class StoreApi extends runtime.BaseAPI {
             method: 'POST',
             headers: headerParameters,
             query: queryParameters,
-            body: OrderToJSON(body),
+            body: OrderToJSON(requestParameters.body),
         });
 
         return new runtime.JSONApiResponse(response, (jsonValue) => OrderFromJSON(jsonValue));
     }
 
-    /**
-     * Place an order for a pet
-     */
+   /**
+    * Place an order for a pet
+    */
     async placeOrder(body: Order): Promise<Order> {
-        const response = await this.placeOrderRaw(body);
+        const response = await this.placeOrderRaw({ body: body });
         return await response.value();
     }
